interface/requirements.txt
asgiref==3.5.2
asttokens==2.2.1
attrs==22.2.0
autopep8==2.0.2
backcall==0.2.0
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==2.1.1
click==8.1.3
decorator==5.1.1
Django==4.1.3
django-crispy-forms==2.0
django-debug-toolbar==3.7.0
django-extensions==3.2.1
django-ninja==0.21.0
djangorestframework==3.14.0
exceptiongroup==1.1.0
executing==1.2.0
Faker==12.0.1
gunicorn==20.1.0
idna==3.4
iniconfig==2.0.0
ipython==8.14.0
jedi==0.18.2
kafka==1.3.5
matplotlib-inline==0.1.6
mixer==7.2.2
orjson==3.8.7
packaging==23.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pluggy==1.0.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
pydantic==1.10.5
Pygments==2.15.1
pysqlite3==0.5.0
pytest==7.2.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7.1
requests==2.28.1
shortuuid==1.0.9
simple-print==1.7.5
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
termcolor==1.1.0
tomli==2.0.1
traitlets==5.9.0
typing_extensions==4.4.0
ujson==5.7.0
urllib3==1.26.13
wcwidth==0.2.6
whitenoise==6.2.0
